requirements.txt
alembic==1.9.4
aniso8601==9.0.1
click==8.1.3
colorama==0.4.6
Flask==2.2.3
Flask-Migrate==4.0.4
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
importlib-metadata==6.0.0
importlib-resources==5.12.0
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
packaging==23.0
password-strength==0.0.3.post2
psycopg2==2.9.5
python-decouple==3.7
pytz==2022.7.1
six==1.16.0
SQLAlchemy==2.0.4
SQLAlchemy-Utc==0.14.0
typing_extensions==4.5.0
Werkzeug==2.2.3
zipp==3.14.0
